requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv
pydantic>=2.6.0  # ← Newer version with pre-built wheels
python-multipart
bcrypt
python-jose[cryptography]
alembic

# Feature Extraction
radon
# Skip transformers and torch for now - we'll add them later if needed

# ML & Data Science
scikit-learn
xgboost
# Skip shap for now - we'll use simple feature importance instead
pandas
numpy
joblib

# OpenAI for embeddings & feedback
openai